@@ -33,7 +33,7 @@
                 {
                     if (PostPublicationStatus.Approved != post.PublicationStatus)
                     {
-                        <MudChip Variant=@Variant.Outlined Color=@GetPublicationChipColor(post.PublicationStatus) Size=@Size.Small>
+                        <MudChip T="string" Variant=@Variant.Outlined Color=@GetPublicationChipColor(post.PublicationStatus) Size=@Size.Small>
                             @Enum.GetName(post.PublicationStatus)
                         </MudChip>
                     }
