@@ -1,5 +1,4 @@
 ﻿using IdentityModel;
-using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using StoryBlog.Web.Microservices.Identity.Application.Authorization.Responses;
@@ -16,11 +15,11 @@ namespace StoryBlog.Web.Microservices.Identity.Application.Endpoints.Results;
 
 internal class AuthorizeResult : IEndpointResult
 {
-    private IdentityServerOptions options;
-    private IUserSession userSession;
-    private IMessageStore<ErrorMessage> errorMessageStore;
-    private IServerUrls urls;
-    private ISystemClock clock;
+    private IdentityServerOptions? options;
+    private IUserSession? userSession;
+    private IMessageStore<ErrorMessage>? errorMessageStore;
+    private IServerUrls? urls;
+    private TimeProvider? clock;
 
     public AuthorizeResponse Response
     {
@@ -38,7 +37,7 @@ internal class AuthorizeResult : IEndpointResult
         IUserSession userSession,
         IMessageStore<ErrorMessage> errorMessageStore,
         IServerUrls urls,
-        ISystemClock clock)
+        TimeProvider clock)
         : this(response)
     {
         this.options = options;
@@ -68,7 +67,7 @@ internal class AuthorizeResult : IEndpointResult
         {
             // success response -- track client authorization for sign-out
             //_logger.LogDebug("Adding client {0} to client list cookie for subject {1}", request.ClientId, request.Subject.GetSubjectId());
-            await userSession.AddClientIdAsync(Response.Request.ClientId);
+            await userSession!.AddClientIdAsync(Response.Request!.ClientId!);
         }
 
         await RenderAuthorizeResponseAsync(context);
@@ -103,19 +102,19 @@ internal class AuthorizeResult : IEndpointResult
         options ??= context.RequestServices.GetRequiredService<IdentityServerOptions>();
         userSession ??= context.RequestServices.GetRequiredService<IUserSession>();
         errorMessageStore ??= context.RequestServices.GetRequiredService<IMessageStore<ErrorMessage>>();
-        urls = urls ?? context.RequestServices.GetRequiredService<IServerUrls>();
-        clock ??= context.RequestServices.GetRequiredService<ISystemClock>();
+        urls ??= context.RequestServices.GetRequiredService<IServerUrls>();
+        clock ??= context.RequestServices.GetRequiredService<TimeProvider>();
     }
 
     private async Task RenderAuthorizeResponseAsync(HttpContext context)
     {
-        if (OidcConstants.ResponseModes.Query == Response.Request.ResponseMode ||
-            OidcConstants.ResponseModes.Fragment == Response.Request.ResponseMode)
+        if (OidcConstants.ResponseModes.Query == Response.Request?.ResponseMode ||
+            OidcConstants.ResponseModes.Fragment == Response.Request?.ResponseMode)
         {
             context.Response.SetNoCache();
             context.Response.Redirect(BuildRedirectUri());
         }
-        else if (Response.Request.ResponseMode == OidcConstants.ResponseModes.FormPost)
+        else if (Response.Request?.ResponseMode == OidcConstants.ResponseModes.FormPost)
         {
             context.Response.SetNoCache();
             AddSecurityHeaders(context);
@@ -130,11 +129,11 @@ internal class AuthorizeResult : IEndpointResult
 
     private void AddSecurityHeaders(HttpContext context)
     {
-        context.Response.AddScriptCspHeaders(options.Csp, "sha256-orD0/VhH8hLqrLxKHD/HUEMdwqX6/0ve7c5hspX5VJ8=");
+        context.Response.AddScriptCspHeaders(options!.Csp, "sha256-orD0/VhH8hLqrLxKHD/HUEMdwqX6/0ve7c5hspX5VJ8=");
 
-        if (false == context.Response.Headers.ContainsKey(Core.HeaderNames.ReferrerPolicy))
+        if (false == context.Response.Headers.ContainsKey(HeaderNames.ReferrerPolicy))
         {
-            context.Response.Headers.Add(Core.HeaderNames.ReferrerPolicy, "no-referrer");
+            context.Response.Headers.Append(HeaderNames.ReferrerPolicy, "no-referrer");
         }
     }
 
@@ -143,14 +142,9 @@ internal class AuthorizeResult : IEndpointResult
         var uri = Response.RedirectUri;
         var query = Response.ToNameValueCollection().ToQueryString();
 
-        if (Response.Request.ResponseMode == OidcConstants.ResponseModes.Query)
-        {
-            uri = uri.AddQueryString(query);
-        }
-        else
-        {
-            uri = uri.AddHashFragment(query);
-        }
+        uri = OidcConstants.ResponseModes.Query == Response.Request?.ResponseMode
+            ? uri!.AddQueryString(query)
+            : uri!.AddHashFragment(query);
 
         if (Response.IsError && !uri.Contains(UriSymbols.HashMark))
         {
@@ -165,10 +159,14 @@ internal class AuthorizeResult : IEndpointResult
     {
         var html = FormPostHtml;
 
-        var url = Response.Request.RedirectUri;
-        url = HtmlEncoder.Default.Encode(url);
-        html = html.Replace("{uri}", url);
-        html = html.Replace("{body}", Response.ToNameValueCollection().ToFormPost());
+        var url = Response.Request?.RedirectUri;
+        
+        if (null != url)
+        {
+            url = HtmlEncoder.Default.Encode(url);
+            html = html.Replace("{uri}", url);
+            html = html.Replace("{body}", Response.ToNameValueCollection().ToFormPost());
+        }
 
         return html;
     }
@@ -178,28 +176,28 @@ internal class AuthorizeResult : IEndpointResult
         var errorModel = new ErrorMessage
         {
             RequestId = context.TraceIdentifier,
-            Error = Response.Error,
-            ErrorDescription = Response.ErrorDescription,
-            UiLocales = Response.Request.UiLocales,
+            Error = Response.Error!,
+            ErrorDescription = Response.ErrorDescription!,
+            UiLocales = Response.Request!.UiLocales,
             DisplayMode = Response.Request.DisplayMode,
-            ClientId = Response.Request.ClientId
+            ClientId = Response.Request.ClientId!
         };
 
-        if (null != Response.RedirectUri && null != Response.Request.ResponseMode)
+        if (Response is { RedirectUri: not null, Request.ResponseMode: not null })
         {
             // if we have a valid redirect uri, then include it to the error page
             errorModel.RedirectUri = BuildRedirectUri();
             errorModel.ResponseMode = Response.Request.ResponseMode;
         }
 
-        var message = new Message<ErrorMessage>(errorModel, clock.UtcNow.UtcDateTime);
-        var id = await errorMessageStore.WriteAsync(message);
-        var errorUrl = options.UserInteraction.ErrorUrl;
-        var url = errorUrl?.AddQueryString(options.UserInteraction.ErrorIdParameter, id);
+        var message = new Message<ErrorMessage>(errorModel, clock!.GetUtcNow().UtcDateTime);
+        var id = await errorMessageStore!.WriteAsync(message);
+        var errorUrl = options!.UserInteraction.ErrorUrl;
+        var url = errorUrl?.AddQueryString(options!.UserInteraction.ErrorIdParameter!, id);
 
-        context.Response.Redirect(urls.GetAbsoluteUrl(url));
+        context.Response.Redirect(urls!.GetAbsoluteUrl(url));
     }
-
+    
     #region Private
 
     private const string FormPostHtml = "<html><head><meta http-equiv='X-UA-Compatible' content='IE=edge' /><base target='_self'/></head><body><form method='post' action='{uri}'>{body}<noscript><button>Click to continue</button></noscript></form><script>window.addEventListener('load', function(){document.forms[0].submit();});</script></body></html>";
